+++ packages/brick_sqlite_generators/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+## 1.1.4
+
 * Fix casting when deserializing enums. The analyzer does not alert for `cast` on a list that could contain nullable values; if the field type is non-nullable, null types must be removed before the cast.
 * Remove unnecessary import in `schema.g.dart`